fixed conflict when to find the nearest screening schedule by adding selection condition auditorium.id=? (ScreeningRepository) config bookingticket-service complement in: ScreeningService SeatService SeatReservationService

+++ auditorium/src/main/java/com/whatamovie/auditorium/repository/SeatReservationRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface SeatReservationRepository extends JpaRepository<SeatReservation, Integer> {
+public interface SeatReservationRepository extends JpaRepository<SeatReservation, Long> {
     @Query(value = "select * from seat_reservation where screening_id=?1 and (select auditorium_id from seat where seat.id=seat_reservation.seat_id)=?2",nativeQuery = true)
     List<SeatReservation> findAllByScreeningAndAuditorium(long screeningId, long auditoriumId);
     @Query(value = "select * from seat_reservation where seat_id=?3 and screening_id=?1 (select auditorium_id from seat where seat.id=seat_reservation.seat_id)=?2",nativeQuery = true)
